+++ android/versioned-abis/expoview-abi39_0_0/src/main/java/abi39_0_0/expo/modules/notifications/notifications/categories/ExpoNotificationCategoriesModule.java
@@ -12,6 +12,7 @@ import abi39_0_0.org.unimodules.core.interfaces.ExpoMethod;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
@@ -30,7 +31,6 @@ public class ExpoNotificationCategoriesModule extends ExportedModule {
   private static final String OPTIONS_KEY = "options";
   private static final String OPENS_APP_TO_FOREGROUND_KEY = "opensAppToForeground";
   private static final String TEXT_INPUT_OPTIONS_KEY = "textInput";
-  private static final String SUBMIT_BUTTON_TITLE_KEY = "submitButtonTitle";
   private static final String PLACEHOLDER_KEY = "placeholder";
 
   private final NotificationsHelper mNotificationsHelper;
@@ -66,11 +66,11 @@ public class ExpoNotificationCategoriesModule extends ExportedModule {
     List<NotificationAction> actions = new ArrayList();
     for (HashMap<String, Object> actionMap : actionArguments) {
       MapArguments actionParams = new MapArguments(actionMap);
-      MapArguments actionOptions = new MapArguments(actionParams.getMap(OPTIONS_KEY));
+      MapArguments actionOptions = new MapArguments(actionParams.getMap(OPTIONS_KEY, Collections.emptyMap()));
       MapArguments textInputOptions = actionParams.containsKey(TEXT_INPUT_OPTIONS_KEY) ? new MapArguments(actionParams.getMap(TEXT_INPUT_OPTIONS_KEY)) : null;
       if (textInputOptions != null) {
         actions.add(new TextInputNotificationAction(actionParams.getString(IDENTIFIER_KEY, null), actionParams.getString(BUTTON_TITLE_KEY, null),
-          actionOptions.getBoolean(OPENS_APP_TO_FOREGROUND_KEY, true), textInputOptions.getString(SUBMIT_BUTTON_TITLE_KEY, null), textInputOptions.getString(PLACEHOLDER_KEY, null)));
+          actionOptions.getBoolean(OPENS_APP_TO_FOREGROUND_KEY, true), textInputOptions.getString(PLACEHOLDER_KEY, null)));
       } else {
         actions.add(new NotificationAction(actionParams.getString(IDENTIFIER_KEY, null), actionParams.getString(BUTTON_TITLE_KEY, null), actionOptions.getBoolean(OPENS_APP_TO_FOREGROUND_KEY, true)));
       }